Alway call the memory allocation function with the correct type for the first argument. The problem is that on some OSes the iovec struct is not POSIX complian, the iov_len is not a size_t but simply an int. This patch, add a local variable (type size_t) to use with the memory allocation function, and then put back the value in the iov_len field.
This commit was SVN r7615.

+++ ompi/datatype/dt_unpack.c
@@ -333,7 +333,7 @@ static int ompi_convertor_unpack_homogeneous_contig( ompi_convertor_t* pConv,
     for( iov_count = 0; iov_count < (*out_size); iov_count++ ) {
         pSrcBuf = (char*)iov[iov_count].iov_base;
         remaining = pConv->count * pData->size - pConv->bConverted;
-        if( remaining > iov[iov_count].iov_len )
+        if( remaining > (uint32_t)iov[iov_count].iov_len )
             remaining = iov[iov_count].iov_len;
         bConverted = remaining; /* how much will get unpacked this time */
